requirements.txt
bidict==0.23.1
blinker==1.9.0
blobfile==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.5.1
fsspec==2025.2.0
h11==0.14.0
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
pycryptodomex==3.21.0
python-docx==1.1.2
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
sacremoses==0.1.1
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==75.8.0
simple-websocket==1.1.0
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
